Moved wrappers in util class
